[message] Move tests internally and make private functions static
