--- test/message/main.c
@@ -1,40 +1,0 @@
-/**
- * Copyright (C) 2020-2022 Vito Tardia <https://vito.tardia.me>
- *
- * This file is part of C2Hat
- *
- * C2Hat is a simple client/server TCP chat written in C
- *
- * C2Hat is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * C2Hat is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with C2Hat. If not, see <https://www.gnu.org/licenses/>.
- */
-
-#include <stdio.h>
-#include <string.h>
-
-#include "message/message.h"
-#include "message_tests.h"
-
-int main() {
-
-  // TestMessage_getType();
-  TestMessage_format();
-  TestMessage_getUser();
-  TestMessage_get();
-  TestC2HMessage_get();
-  TestC2HMessage_create();
-  TestC2HMessage_createFromString();
-
-  printf("\n");
-  return 0;
-}
