--- test/message/message_tests.h
@@ -1,32 +1,0 @@
-/**
- * Copyright (C) 2020-2022 Vito Tardia <https://vito.tardia.me>
- *
- * This file is part of C2Hat
- *
- * C2Hat is a simple client/server TCP chat written in C
- *
- * C2Hat is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * C2Hat is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with C2Hat. If not, see <https://www.gnu.org/licenses/>.
- */
-
-#ifndef MESSAGE_TESTS_H
-#define MESSAGE_TESTS_H
-  void TestMessage_getType();
-  void TestMessage_getContent();
-  void TestMessage_format();
-  void TestMessage_getUser();
-  void TestMessage_get();
-  void TestC2HMessage_get();
-  void TestC2HMessage_create();
-  void TestC2HMessage_createFromString();
-#endif
